requirements.txt
Django==1.9
djangorestframework==3.3.1
pygments
httpie
mod-wsgi==4.4.11
mysqlclient==1.3.6

django-extensions==1.5.2
django-crispy-forms==1.5.2
django-reversion==1.10.2
django-braces==1.8.1
django-filter==0.11.0
django-cacheops==2.4.3
django-pipeline==1.6.8
pystalkd==1.2.1
django-dbbackup==2.4.1

# dev
ipython  # interactive python
readline  # for the python shell to have history
django-debug-toolbar
django-debug-toolbar-template-profiler
django-debug-toolbar-template-timings